LayoutTests/media/video-muted-after-setting-page-muted-state.html

<video controls></video>
<p>Test 'muted' IDL attribute after setting the Page's muted state. The Page's muted state should not affect the 'muted' IDL attribute.<p>
<script src=media-file.js></script>
<script src=video-test.js></script>
<script>
    testExpected("video.muted", false);
    if (window.internals)
        internals.setPageMuted("audio");
    testExpected("video.muted", false);

    run("video.muted = true");
    testExpected("video.muted", true);
    if (window.internals)
        internals.setPageMuted("");
    testExpected("video.muted", true);
    endTest();
</script>
